Move ui components to separate package
