+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme"
         tools:ignore="GoogleAppIndexingWarning">
-        <activity android:name="com.marekmacko.kotlinapp.MainActivity">
+        <activity android:name="com.marekmacko.kotlinapp.ui.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
