+++ app/src/main/java/com/marekmacko/kotlinapp/WeatherComponent.kt
@@ -1,7 +1,9 @@
 package com.marekmacko.kotlinapp
 
 import com.marekmacko.kotlinapp.api.NetworkModule
+import com.marekmacko.kotlinapp.mvp.WeatherPresenterModule
 import com.marekmacko.kotlinapp.repository.RepositoryModule
+import com.marekmacko.kotlinapp.ui.WeeklyForecastFragment
 import dagger.Component
 import javax.inject.Singleton
