+++ app/src/main/res/layout/activity_main.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.marekmacko.kotlinapp.MainActivity">
+    tools:context="com.marekmacko.kotlinapp.ui.MainActivity">
 
 </android.support.constraint.ConstraintLayout>
